Move ReceiptStatusMapper to MarketDocument
