+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/ChargeLinkReceiptBundle/Cim/ChargeLinkReceiptCimSerializer.cs
@@ -20,6 +20,7 @@ using GreenEnergyHub.Charges.Domain.AvailableData;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
 using GreenEnergyHub.Charges.Infrastructure.Cim;
 using GreenEnergyHub.Charges.Infrastructure.Configuration;
+using GreenEnergyHub.Charges.Infrastructure.MarketDocument;
 using GreenEnergyHub.Charges.Infrastructure.MarketDocument.Cim;
 using NodaTime;
